requirements.txt
fastapi
uvicorn
pytesseract
numpy
opencv-python
python-dotenv
pillow
requests
bs4
langchain
openai
tiktoken
